Updated Script
Consolidated List/Folder and List/Folder/Continue calls, added filtering for folders in the initial list creation instead of after (see: alluserfiles). Added flow control for if the user is not the owner of a shared folder to skip it. Removed MassUserCopy() function, I'll add a separate script for that process.